feat(toolkit): Implement human verification for api calling (#343)
## 🚀 Summary

When an agent wants to call an api to create e.g. a Jira ticket it
should ask for verification. This implemenation allow to define a
Verification manager that allows to create the right assistant message,
checks the user message for an indication of a verification, handles the
api call.