chore(unique_toolkit): Migrate failsafe execution handler from agentic tools to unique_toolkit _common (#770)
## 🚀 Summary
Refs: https://unique-ch.atlassian.net/browse/UN-15772

Briefly describe the changes introduced in this PR.
The current execution handler lives in the agentic tools, which are used
in postprocessor, websearch, etc. Moving it to the _common folder in
order for all packages to inherit it from makes much sense. This will
also be propagated to the monolith as well to remove the duplication

## ✅ Changes

- [X] Refactored code / cleaned up

Co-authored-by: Azeez Raheem <raheemazeez4@gmil.com>